+++ test/repositories/RequestRepository/getByMessageHash.test.ts
@@ -2,7 +2,7 @@ import 'mocha';
 
 import BigNumber from 'bignumber.js';
 
-import MessageTransferRequest from '../../../src/models/Request';
+import MessageTransferRequest from '../../../src/models/MessageTransferRequest';
 import Repositories from '../../../src/repositories/Repositories';
 import assert from '../../test_utils/assert';
 import StubData from '../../test_utils/StubData';
