+++ test/repositories/RequestRepository/save.test.ts
@@ -18,7 +18,7 @@ import 'mocha';
 
 import BigNumber from 'bignumber.js';
 
-import MessageTransferRequest from '../../../src/models/Request';
+import MessageTransferRequest from '../../../src/models/MessageTransferRequest';
 import Repositories from '../../../src/repositories/Repositories';
 import assert from '../../test_utils/assert';
 import Util from './util';
