+++ test/test_utils/StubData.ts
@@ -5,7 +5,7 @@ import AuxiliaryChain from '../../src/models/AuxiliaryChain';
 import ContractEntity, { EntityType } from '../../src/models/ContractEntity';
 import Gateway from '../../src/models/Gateway';
 import Message from '../../src/models/Message';
-import MessageTransferRequest from '../../src/models/Request';
+import MessageTransferRequest from '../../src/models/MessageTransferRequest';
 import { GatewayType } from '../../src/repositories/GatewayRepository';
 import {
   MessageDirection, MessageStatus, MessageType,
